refactor(devtui): split setup_guide into model/view/apply files
setup_guide.go was 769 lines mixing model, step transitions, rendering,
and config persistence. Split along those seams:

- setup_guide.go        — model, step transitions
- setup_guide_view.go   — rendering methods
- setup_guide_apply.go  — applyToConfig, ensureDeploymentHelper, localConfig

No behavior changes; existing setup_guide_test.go continues to cover all three.
